Make snapshot writing a policy, not a switch
The record side branched on EpilogueSnapshotType inside the epilogue
lambda, over three static writers with three mismatched signatures, and
"best" mode compared them through computeMnemeBytesSnapshotSize, a
by-hand copy of the bytes layout that had to be edited in lockstep with
the writer.

SnapshotInput collapses the three signatures into one argument.
FormatWriter owns the shape every format shares, a header followed by a
payload, so a format supplies only header() and writePayload(). "Best" is
now a composite writer that measures its two members and delegates to the
smaller, rather than a third policy that re-derives both sizes.

BytesWriter::measure stays analytic, because the default counting pass
would copy every blob device-to-host just to size it. ValidateMeasure in
the unit test now pins both writers to the size they actually produced,
which is the check the hand-maintained estimate never had.

makeEpilogueWriter is the only remaining switch on EpilogueSnapshotType.
MnemeSnapshot is gone.

No on-disk change.
